Make OPON an action and connect it to Ctrl+O
There was already an old comment saying it was connected to Ctrl+O.
